Bug fixes and additions to Epstein Civil Violence
This update changes a small set of bugs that was causing the model to
not exhibit the cyclical explosions of active agents seen in the
original Epstein paper. It also adds some small functionality
to the browser display, introducing sliders to allow the user
to change model parameters on the fly, and watch a graph
of agent state numbers.

Bug Fixes
- Fix agent state logic
- Fix Moore vs Von Neuman neighborhood selection to use
    Epstein paper behavior
- Fix agent vision bug (was 1 and not parameter `vision` when
    calculating neighborhood, which also fixes agent movement issue)
- Fix arrestee state issue (arrestee was not being reset to "Quiescent",
    and so was counting as "Active" for all nearby agents)

Additions and Changes
- add model attributes to count agents and display a graph of agent
    states
- Change display of agents where jailed agents are represented by grey
    square box
- Added sliders for model attributes